Truncate queryset results more than 10
With large queryset result sets for a versioned field, the versioning UI becomes unresponsive. This has been pinpointed to codemirror having to create an instance for each result in the result set which causes the page to break. Assuming that users are not going to be looking at each result in large result sets, we now truncate those result sets to only show 10 items and inform the user about how many were truncated